neomutt: 20211015
Turns out we can make the manpages without an internet connection.
this ./configures as if it would build the docs, so that the Makefile contains the logic
required to build the manpages, but then the all-docs target is removed from ALL_TARGETS.
We later call `make` separately to build just what we want.